silson_rag/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
pytest>=7.0.0
